Fetch notes and add get-notes case to the reducer
Add useEffect in the ToDoList component in order to fetch the notes from the backend before the component is rendered. As the notes fetching triggers the dispatch (case get-notes) the context state gets updated and the component ToDoList render the notes stored in this new state of the context. Also, some descriptive comments were added along the code.