Save note on the DB from the frontend
In the function OnAdd, triggered by the button event, the POST request was created in order to save a new note in the DB. For this purpose, the onAdd function was defined as asynchronus as the fetch method returs a promise. In the fetch function the body, the headers and the method of the request (POST) were sent as an object. Right after, the dispatch triggers the reducer case 'add-note', updating the context state in the frontend.