Update note on DB from UI
In the onCheck event of the ToDoList component it was made a fetch request to the backend specifying the PUT method. To do this, the done property of a note was changed by using the spread operator, then sending this object through the body of the fetch request. Here, the promises were handled by both the async/await approach and the .then() approach. After that, the object that is returned by the fetch is sent through the payload in the dispath, that triggers the update-note case in the reducer.